WIP: test react-window to speed up init of authors panel.
Results: it works, the panel displays faster in Firefox (the
only browser that was slow in my tests), and the added 2kb
are negligible.
However, the documentation of ReactWindow is lacking, and I
can't figure out how to scroll to a given index of the list,
in order to implement the index feature.
I'm giving up on this for now, because I reckon we should
replace the list by an autocomplete anyways, at least on
desktop.